refactor: style 관련 color 계산 함수 util로 분리
